requirements.txt
django
djangorestframework
djongo==1.3.6
pymongo[srv]